backend/requirements.txt
fastapi==0.135.2
uvicorn[standard]==0.42.0
sqlalchemy[asyncio]==2.0.48
asyncpg==0.31.0
alembic==1.18.4
pydantic==2.12.5
pydantic-settings==2.13.1
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22
stripe==14.4.1
azure-ai-formrecognizer==3.3.3
aiofiles==25.1.0
boto3==1.42.74
celery[redis]==5.6.2
httpx==0.28.1
